+++ libs/api/notifications/service/feature/src/lib/events/send-mail.event.spec.ts
@@ -1,4 +1,4 @@
-import { ApiNotificationsService } from '../../api-notifications-service-feature.service';
+import { ApiNotificationsService } from '../api-notifications-service-feature.service';
 import { SendMailEvent } from './send-mail.event';
 
 describe('SendMailEvent', () => {
