+++ libs/api/notifications/service/feature/src/lib/events/send-mail.event.ts
@@ -1,4 +1,4 @@
-import { ApiNotificationsService } from "../../api-notifications-service-feature.service";
+import { ApiNotificationsService } from "../api-notifications-service-feature.service";
 export class SendMailEvent {
     constructor(private readonly data:ApiNotificationsService){
         //we set the data that we will use here waiting for repository(API)
