+++ libs/api/notifications/service/feature/src/lib/events/send-mail.handler.spec.ts
