+++ libs/api/notifications/service/feature/src/lib/queries/get-all-user-notifications.handlers.ts
@@ -1,7 +1,7 @@
 
 import { NotificationsRepository } from "@graduates/api/notifications/repository/data-access";
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
-import { GetAllUserNotificationsQuery } from '../impl/get-all-user-notifications.query';
+import { GetAllUserNotificationsQuery } from './get-all-user-notifications.query';
 // import { ApiNotificationsService } from '../../api-notifications-service-feature.service';
 
 
@@ -11,6 +11,15 @@ export class GetAllUserNotificationsHandler implements IQueryHandler<GetAllUserN
 
   async execute() {
     return this.repository.findNotificationsAll();
+  }
+}
+
+@QueryHandler(GetUserNotificationsReceivedQuery)
+export class GetUserNotificationsReceivedQuery implements IQueryHandler<GetUserNotificationsReceivedQuery> {
+  constructor(private readonly repository: NotificationsRepository) {}
+
+  async execute() {
+    // return this.repository.findNotificationById();
   }
 }
